remove 0-main and print_array files, editted the 0-bubble_sort file and added helper_func file of the helper functions
